[REFACTOR] : input을 초기화 하는 로직을 resetInput을 사용하도록 수정
