+++ src/event/reset/reset.js
@@ -2,6 +2,7 @@ import lottoStore from "../../store/lottoStore.js";
 import priceStore from "../../store/priceStore.js";
 import setPurchaseDetailVisibility from "../setPurchaseDetailVisibility.js";
 import resetWinningHistoryUI from "./resetWinningHistoryUI.js";
+import resetInput from "../../utils/resetInput.js";
 
 const reset = () => {
   priceStore.setPrice(0);
@@ -14,11 +15,11 @@ const reset = () => {
   const bonusNumberInput = document.querySelector(".bonusNumberInput");
 
   winningNumberInputs.forEach((input) => {
-    input.value = "";
+    resetInput(input);
   });
+  resetInput(bonusNumberInput);
 
   resetWinningHistoryUI();
-  bonusNumberInput.value = "";
   purchasedLottos.textContent = "";
   setPurchaseDetailVisibility("off");
 };
